Fixed ability handling by making it more explainable.
Ability + Tile => Ability Effect + AbilityEffectMarker

propagation of effects now only works for the same ability effect and that is checked via the target tile's list of declared ability effectors. This way a burning tree does not propagate the generation of lava when the two effectors happen to sit on neighbouring tiles.